move class methods to a module
# Conflicts:
#	ruby/lib/ccxt/base/exchange.rb
